Start implementing sysfs
Note, that this is different from sysfd (This one's about filenames and
such).
Meant as a seamless interface between vfs/usercode and the host OS, the
new FS API is planned to feature interchangable functions that
enable/disable the VFS, as well as automatic path formating to expand
env vars and the home prefix.